requirements.txt


# Core
discord.py>=2.3,<3.0
python-dotenv
aiohttp>=3.8,<4.0
aiosqlite>=0.19,<1.0
requests
httpx
aiofiles

# AI & Machine Learning
torch>=2.0.0
transformers>=4.48.0
accelerate>=0.20.0
openai
google-generativeai
numpy>=1.23,<2.0
matplotlib>=3.7.0

# Voice & Audio
PyNaCl>=1.5,<2.0
git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git
edge-tts
gTTS
openai-whisper>=20230314.0
yt-dlp>=2024.3.10
soundfile>=0.12.0

# Vision & Image
pillow>=9.5,<10.0
pytesseract>=0.3,<1.0
opencv-python-headless>=4.8,<5.0
mss

# Search & Google APIs
google-search-results>=2.0,<3.0
duckduckgo-search>=4.0,<7.0
google-api-python-client
google-cloud-vision
google-auth-oauthlib
google-auth
firebase-admin
beautifulsoup4

# System
psutil
pycaw; sys_platform == 'win32'
comtypes; sys_platform == 'win32'
pycountry>=22.3,<24.0
websocket-client
pytz

# Web API (Dashboard/IPC)
fastapi
uvicorn[standard]
python-multipart

# Development & Testing (Local only)
pytest
pytest-asyncio
ruff
mypy
pip-tools
chromadb==0.6.3
playwright
